refactor: centralize buildUrl helper
