+++ frontend/src/components/books/BookCard.js
@@ -1,18 +1,7 @@
 import Link from "next/link";
 import { useTranslation } from "next-i18next";
 import { FiEye, FiBookOpen, FiEdit, FiTrash2 } from "react-icons/fi";
-
-const API_BASE = process.env.NEXT_PUBLIC_API_BASE_URL || "";
-
-const buildUrl = (path) => {
-  if (!path) return null;
-  if (/^https?:/i.test(path)) return path;
-  const uploadsIndex = path.indexOf("/uploads");
-  const relative = uploadsIndex !== -1 ? path.substring(uploadsIndex) : path;
-  const normalized = relative.startsWith("/") ? relative : `/${relative}`;
-  return `${API_BASE}${normalized}`;
-};
-
+import { buildUrl } from "@/utils/url";
 
 export default function BookCard({
   book,
