+++ frontend/src/pages/dashboard/admin/books/index.js
@@ -14,16 +14,7 @@ import { useTranslation } from "next-i18next";
 import { FiPlus, FiSearch, FiTrash2, FiChevronLeft, FiChevronRight, FiFilter, FiX, FiEdit, FiEye } from "react-icons/fi";
 import { Switch } from '@headlessui/react';
 import ConfirmModal from "@/components/common/ConfirmModal";
-
-const API_BASE = process.env.NEXT_PUBLIC_API_BASE_URL || "";
-const buildUrl = (path) => {
-  if (!path) return null;
-  if (/^https?:/i.test(path)) return path;
-  const uploadsIndex = path.indexOf("/uploads");
-  const relative = uploadsIndex !== -1 ? path.substring(uploadsIndex) : path;
-  const normalized = relative.startsWith("/") ? relative : `/${relative}`;
-  return `${API_BASE}${normalized}`;
-};
+import { buildUrl } from "@/utils/url";
 
 function AdminBooksPage() {
   const { t } = useTranslation("dashboard");
