+++ frontend/src/services/bookService.js
@@ -1,15 +1,5 @@
 import api from "@/services/api/api";
-
-const API_BASE = process.env.NEXT_PUBLIC_API_BASE_URL || "";
-
-const buildUrl = (path) => {
-  if (!path) return null;
-  if (/^https?:/i.test(path)) return path;
-  const uploadsIndex = path.indexOf("/uploads");
-  const relative = uploadsIndex !== -1 ? path.substring(uploadsIndex) : path;
-  const normalized = relative.startsWith("/") ? relative : `/${relative}`;
-  return `${API_BASE}${normalized}`;
-};
+import { buildUrl } from "@/utils/url";
 
 const formatBook = (book) => ({
   ...book,
